單價分析表數量可使用小數

+++ database/migrations/2015_12_13_051109_create_work_items_table.php
@@ -22,7 +22,7 @@ class CreateWorkItemsTable extends Migration
             $table->string('name');
 
             $table->integer('order')->unsigned();
-            $table->integer('amount')->unsigned();
+            $table->double('amount', 15, 8)->unsigned();
             $table->integer('unit_price')->unsigned();
 
             $table->timestamps();
